// frontend/src/lib/api/model/lobbyFiltersStatus.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Student Arena API
 * API Documentation
 * OpenAPI spec version: 1.0
 */

/**
 * Filter lobbies by status
 */
export type LobbyFiltersStatus = typeof LobbyFiltersStatus[keyof typeof LobbyFiltersStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LobbyFiltersStatus = {
  SCHEDULED: 'SCHEDULED',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
} as const;