// frontend/src/lib/api/model/lobbyFiltersMatchType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Student Arena API
 * API Documentation
 * OpenAPI spec version: 1.0
 */

/**
 * Filter lobbies by match type
 */
export type LobbyFiltersMatchType = typeof LobbyFiltersMatchType[keyof typeof LobbyFiltersMatchType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LobbyFiltersMatchType = {
  Queue: 'Queue',
  Team: 'Team',
} as const;